Lab3 implementation v2.0

+++ src/pl/gdynia/amw/oop/Lab3/Square.java
@@ -2,7 +2,7 @@ package pl.gdynia.amw.oop.Lab3;
 
 public class Square extends Figure {
 
-    private int length;
+    private Float length;
 
     public Square() {
 
@@ -11,7 +11,7 @@ public class Square extends Figure {
     @Override
     public void printSurfaceArea() {
         if (length > 0) {
-            System.out.printf("Surface area of the square = %d%n", length * length);
+            System.out.printf("Surface area of the square = %f%n", length * length);
         } else {
             System.out.println("Value of side of a Square should be a positive number");
         }
@@ -20,17 +20,17 @@ public class Square extends Figure {
     @Override
     public void printCircuit() {
         if (length > 0) {
-            System.out.printf("Circuit of the square = %d%n", 4 * length);
+            System.out.printf("Circuit of the square = %f%n", 4 * length);
         } else {
             System.out.println("Value of side of a Square should be a positive number");
         }
     }
 
-    public int getLength() {
+    public float getLength() {
         return length;
     }
 
-    public void setLength(int length) {
+    public void setLength(float length) {
         this.length = length;
     }
 }
